requirements.txt
# Core dependencies
streamlit>=1.29.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# OCR and document processing
pdfplumber>=0.10.0
PyMuPDF>=1.23.0
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python>=4.8.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# API client
openai>=1.0.0

# Utilities
requests>=2.31.0
regex>=2023.0.0
